cloud_functions/qa_agent/requirements.txt
functions-framework==3.*
flask==3.*
requests==2.31.0
google-cloud-firestore==2.14.0
firebase-admin==6.4.0
google-cloud-secret-manager>=2.0.0
google-generativeai>=0.8.0
